chore: restructure app by moving files into [lng] directory feat: installed necessary dependencies for i18n
