+++ app/[lng]/_components/Map/Map.module.css
