+++ app/[lng]/_components/Map/index.ts
