+++ app/[lng]/_components/NavBar/NavBar.tsx
