+++ app/[lng]/_components/NavBar/index.ts
