+++ app/[lng]/_components/OmamoriCard/OmamoriCard.tsx
