+++ app/[lng]/_components/OmamoriCard/index.ts
