+++ app/[lng]/_fonts/NotoSans-Regular.ttf
