+++ app/[lng]/globals.css
