+++ app/[lng]/login/page.tsx
@@ -1,8 +1,9 @@
 "use client";
+import Link from "next/link";
 import { FormEvent } from "react";
 
-export default function Signup() {
-  // Handle signup
+export default function Login() {
+  // Handle login
   const handleSubmit = (e: FormEvent) => {
     e.preventDefault();
     return;
@@ -14,7 +15,7 @@ export default function Signup() {
         <div className="flex flex-col w-full">
           <div>
             <label
-              htmlFor="signup-email-field"
+              htmlFor="login-email-field"
               className="block mb-2 leading-none"
             >
               Email:
@@ -24,14 +25,14 @@ export default function Signup() {
               type="email"
               aria-label="email"
               placeholder="name@domain.com"
-              data-test="signup-email-field"
+              data-test="login-email-field"
               autoCorrect="off"
               className="w-full h-10 p-2 mb-6 border border-solid rounded focus:outline focus:outline-2 focus:outline-offset-2"
             />
           </div>
           <div>
             <label
-              htmlFor="signup-password-field"
+              htmlFor="login-password-field"
               className="block mb-2 leading-none"
             >
               Password:
@@ -41,39 +42,22 @@ export default function Signup() {
               type="password"
               aria-label="password"
               placeholder="*****"
-              data-test="signup-password-field"
-              autoCorrect="off"
-              className="w-full h-10 p-2 mb-6 border border-solid rounded focus:outline focus:outline-2 focus:outline-offset-2"
-            />
-          </div>
-          <div>
-            <label
-              htmlFor="signup-username-field"
-              className="block mb-2 leading-none"
-            >
-              Username:
-            </label>
-            <input
-              required
-              type="username"
-              aria-label="username"
-              placeholder="omamori"
-              data-test="signup-username-field"
+              data-test="login-password-field"
               autoCorrect="off"
               className="w-full h-10 p-2 mb-6 border border-solid rounded focus:outline focus:outline-2 focus:outline-offset-2"
             />
           </div>
           <div>
             <button className="w-full bg-[var(--primary-button)] rounded-md py-2 px-4">
-              <div className="text-center">Sign up</div>
+              <div className="text-center">Log in</div>
             </button>
           </div>
         </div>
       </form>
       <div>
-        <a href="/login">
-          <p>Have an account? Login</p>
-        </a>
+        <Link href="/signup">
+          <p>Sign up</p>
+        </Link>
       </div>
     </div>
   );
