+++ package.json
@@ -17,9 +17,15 @@
     ]
   },
   "dependencies": {
+    "accept-language": "^3.0.20",
+    "i18next": "^24.2.2",
+    "i18next-browser-languagedetector": "^8.0.4",
+    "i18next-resources-to-backend": "^1.2.1",
     "next": "15.0.1",
     "react": "^18.0.0",
+    "react-cookie": "^7.2.2",
     "react-dom": "^18.0.0",
+    "react-i18next": "^15.4.1",
     "zustand": "^5.0.1"
   },
   "devDependencies": {
